Retrieve tag from git/refs/tags instead repo/tags
Github doesn’t sort tags by created date or even doesn’t allow us to sort tags at all.
So we need to retrieve all tags to be sure we don’t miss a new one that’ll be in the bottom of the list from repo/tags.

This is the case (at least) of `ampproject/amphtml`.

See http://stackoverflow.com/q/19452244/569101