Add Github Client Discovery
Find the best token to use to fetch data to avoid Rate Limit.

The goal is to find the best authenticated method to avoid hitting the Github rate limit.
We first try with the default application authentication (which might work 99% of the time).
And if it fails, we'll try each user until we find one with enough rate limit.
In fact, the more user in database, the bigger chance to never hit the rate limit (this is theorical).

This mostly to handle new users.
When a new user logged in for the first time we need to retrieve all releases/tags of its starred repos and it use a lot (really a lot) of rate limit.

Also, using "ocramius/proxy-manager" to lazy services which require the Github/Client because we don't want the Github Client Discovery to be triggered when clearing cache.