Retrieve release using RSS
Cons:
- we can only retrieve 10 releases
- no caching like “modified since” on the GitHub API

Pros:
- we don’t need to convert the markdown field (already done)
- we won’t hit the rate limit
- only one HTTP request to retrieve all information

Workaround for the first _cons_, introducing a new queue `banditore.sync_versions_info`. As soon as a new version is created when using the RSS method, a message will be posted in that queue to use the API to retrieve information not provided by RSS (for now, only the `prerelease` information).

Also, when fetching version from RSS, if we create 10 new versions it means we might have reached all release from the feed but maybe there is no more to fetch. Queuing a message to use the GitHub API to fetch version will handle them.

Also improving displaying a bit.